+++ app/detail/Shared.tsx
