+++ ppl/detail/Correlation.tsx
@@ -4,7 +4,7 @@ import React, {memo, useMemo} from "react"
 import {reactElementProps} from "src/js/test/integration"
 import LogDetails from "src/js/state/LogDetails"
 import {zng} from "zealot"
-import {Caption, ChartWrap, TableWrap} from "app/detail/Spacing"
+import {Caption, ChartWrap, TableWrap} from "app/detail/Shared"
 import PanelHeading from "app/detail/PanelHeading"
 import EventTimeline from "ppl/detail/EventTimeline"
 import {BrimEvent} from "ppl/detail/models/BrimEvent"
