+++ ppl/detail/RelatedConns.tsx
@@ -3,7 +3,7 @@ import {BrimEvent, BrimEventInterface} from "ppl/detail/models/BrimEvent"
 import React, {memo, useMemo} from "react"
 import brim from "src/js/brim"
 import EventTimeline from "./EventTimeline"
-import {TableWrap, ChartWrap, Caption} from "app/detail/Spacing"
+import {TableWrap, ChartWrap, Caption} from "app/detail/Shared"
 import formatDur from "./util/formatDur"
 import PanelHeading from "app/detail/PanelHeading"
 import Panel from "app/detail/Panel"
